requirements.txt
ansicon==1.89.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blessed==1.21.0
certifi==2025.6.15
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
coverage==7.9.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dnslib==0.9.25
enlighten==1.14.1
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
fqdn==1.5.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.5.0
iniconfig==2.1.0
instaloctrack @ file:///C:/Users/Dell/instaloctrack
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jinxed==1.3.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.24.3
opencv-python==4.10.0.84
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.6.0
pprintpp==0.4.0
prefixed==0.9.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pycountry==24.6.1
pycountry-convert==0.7.2
pycparser==2.22
Pygments==2.19.1
PyMuPDF==1.26.3
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-json-logger==3.2.1
python-pptx==1.0.2
pytz==2024.2
pywin32==305
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
repoze.lru==0.7
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.22.3
selenium==4.34.0
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tk==0.1.0
tomli==2.2.1
tornado==6.4.2
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.14.0
tzdata==2024.2
uri-template==1.3.0
urllib3==2.4.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wsproto==1.2.0
xlsxwriter==3.2.5
zipp==3.20.2
